pyproject.toml
[project]
name = "dormai"
version = "0.1.0"
description = ""
authors = [
    {name = "George Kasparyants",email = "gg_plastt@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
